docker: Refactoring - Move get_docker_container_cpu to separate file
